+++ app/macos/Sources/main.m
@@ -510,7 +510,7 @@
         @"nativeEvents": @(nativeEvents),
         @"installed": @(installed),
         @"appPath": appPath ?: @"",
-        @"version": @"0.8.1",
+        @"version": @"0.8.2",
     }
             function:@"receiveCapabilities"];
 }